feat: support grammar injection
