+++ packages/shikiji/scripts/prepare/constants.ts
@@ -1,0 +1,5 @@
+export const COMMENT_HEAD = `
+/**
+ * Generated by scripts/prepare.ts
+ */
+`.trim()
