+++ packages/shikiji/test/index.test.ts
@@ -145,6 +145,10 @@ describe('should', () => {
           "typescript",
           "vb",
           "vue",
+          "vue-directives",
+          "vue-injection-markdown",
+          "vue-interpolations",
+          "vue-sfc-style-variable-injection",
           "xml",
           "xsl",
           "yaml",
